Moved init and shaderpack code
