+++ src/render_engine/vulkan/vulkan_render_engine.cpp
@@ -25,108 +25,7 @@
 #endif
 
 namespace nova {
-    vulkan_render_engine::vulkan_render_engine(const nova_settings& settings, ttl::task_scheduler* task_scheduler) : render_engine(settings, task_scheduler) {
-        NOVA_LOG(INFO) << "Initializing Vulkan rendering";
 
-        const settings_options& options = settings.get_options();
-
-        validate_mesh_options(options.vertex_memory_settings);
-
-        const auto& version = options.vulkan.application_version;
-
-        VkApplicationInfo application_info;
-        application_info.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-        application_info.pNext = nullptr;
-        application_info.pApplicationName = options.vulkan.application_name.c_str();
-        application_info.applicationVersion = VK_MAKE_VERSION(version.major, version.minor, version.patch);
-        application_info.pEngineName = "Nova renderer 0.1";
-        application_info.apiVersion = VK_API_VERSION_1_1;
-
-        VkInstanceCreateInfo create_info;
-        create_info.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
-        create_info.pNext = nullptr;
-        create_info.flags = 0;
-        create_info.pApplicationInfo = &application_info;
-#ifndef NDEBUG
-        enabled_validation_layer_names.push_back("VK_LAYER_LUNARG_standard_validation");
-#endif
-        create_info.enabledLayerCount = static_cast<uint32_t>(enabled_validation_layer_names.size());
-        create_info.ppEnabledLayerNames = enabled_validation_layer_names.data();
-
-        std::vector<const char*> enabled_extension_names;
-        enabled_extension_names.push_back(VK_KHR_SURFACE_EXTENSION_NAME);
-#ifdef NOVA_LINUX
-        enabled_extension_names.push_back(VK_KHR_XLIB_SURFACE_EXTENSION_NAME);
-#elif NOVA_WINDOWS
-        enabled_extension_names.push_back(VK_KHR_WIN32_SURFACE_EXTENSION_NAME);
-#else
-#error Unsupported Operating system
-#endif
-
-#ifndef NDEBUG
-        enabled_extension_names.push_back(VK_EXT_DEBUG_REPORT_EXTENSION_NAME);
-        enabled_extension_names.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-#endif
-        create_info.enabledExtensionCount = static_cast<uint32_t>(enabled_extension_names.size());
-        create_info.ppEnabledExtensionNames = enabled_extension_names.data();
-
-        NOVA_THROW_IF_VK_ERROR(vkCreateInstance(&create_info, nullptr, &vk_instance), render_engine_initialization_exception);
-
-        uint32_t num_extensions;
-        vkEnumerateInstanceExtensionProperties(nullptr, &num_extensions, nullptr);
-        gpu.available_extensions.resize(num_extensions);
-        vkEnumerateInstanceExtensionProperties(nullptr, &num_extensions, gpu.available_extensions.data());
-
-        std::stringstream ss;
-        for(const VkExtensionProperties& props : gpu.available_extensions) {
-            ss << "\t" << props.extensionName << " version " << props.specVersion << "\n";
-        }
-
-        NOVA_LOG(TRACE) << "Supported extensions:\n" << ss.str();
-
-#ifndef NDEBUG
-        vkCreateDebugUtilsMessengerEXT = reinterpret_cast<PFN_vkCreateDebugUtilsMessengerEXT>(vkGetInstanceProcAddr(vk_instance, "vkCreateDebugUtilsMessengerEXT"));
-        vkDestroyDebugReportCallbackEXT = reinterpret_cast<PFN_vkDestroyDebugReportCallbackEXT>(vkGetInstanceProcAddr(vk_instance, "vkDestroyDebugReportCallbackEXT"));
-
-        VkDebugUtilsMessengerCreateInfoEXT debug_create_info;
-        debug_create_info.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
-        debug_create_info.pNext = nullptr;
-        debug_create_info.flags = VK_DEBUG_REPORT_ERROR_BIT_EXT | VK_DEBUG_REPORT_WARNING_BIT_EXT | VK_DEBUG_REPORT_PERFORMANCE_WARNING_BIT_EXT;
-        debug_create_info.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
-        debug_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
-        debug_create_info.pfnUserCallback = reinterpret_cast<PFN_vkDebugUtilsMessengerCallbackEXT>(&debug_report_callback);
-        debug_create_info.pUserData = this;
-
-        NOVA_THROW_IF_VK_ERROR(vkCreateDebugUtilsMessengerEXT(vk_instance, &debug_create_info, nullptr, &debug_callback), render_engine_initialization_exception);
-#endif
-        // First we open the window. This doesn't depend on anything except the VkInstance/ This method also creates
-        // the VkSurfaceKHR we can render to
-        open_window(settings.get_options().window.width, settings.get_options().window.height);
-
-        // Create the device. This depends on both the VkInstance and the VkSurfaceKHR: we need the VkSurfaceKHR to
-        // make sure we find a device that can present to that surface
-        create_device();
-
-        create_per_thread_command_pools();
-
-        // Create the swapchain. This depends on the VkInstance, VkPhysicalDevice, our pre-thread command pools, and
-        // VkSurfaceKHR. This method also fills out a lot of the information in our vk_gpu_info
-        create_swapchain();
-
-        create_memory_allocator();
-        mesh_memory = std::make_unique<compacting_block_allocator>(settings.get_options().vertex_memory_settings, vma_allocator, graphics_family_index, copy_family_index);
-
-        create_global_sync_objects();
-        create_per_thread_descriptor_pools();
-        create_default_samplers();
-
-#ifndef NDEBUG
-        vkSetDebugUtilsObjectNameEXT = reinterpret_cast<PFN_vkSetDebugUtilsObjectNameEXT>(vkGetDeviceProcAddr(device, "vkSetDebugUtilsObjectNameEXT"));
-        if(vkSetDebugUtilsObjectNameEXT == nullptr) {
-            NOVA_LOG(ERROR) << "Could not load the debug name function";
-        }
-#endif
-    }
 
     vulkan_render_engine::~vulkan_render_engine() {
         vkDeviceWaitIdle(device);
@@ -212,169 +111,6 @@ namespace nova {
         dynamic_textures_need_to_transition = false;
     }
 
-    void vulkan_render_engine::create_swapchain() {
-        NOVA_THROW_IF_VK_ERROR(vkGetPhysicalDeviceSurfaceCapabilitiesKHR(gpu.phys_device, surface, &gpu.surface_capabilities), render_engine_initialization_exception);
-
-        uint32_t num_surface_formats;
-        NOVA_THROW_IF_VK_ERROR(vkGetPhysicalDeviceSurfaceFormatsKHR(gpu.phys_device, surface, &num_surface_formats, nullptr), render_engine_initialization_exception);
-        gpu.surface_formats.resize(num_surface_formats);
-        NOVA_THROW_IF_VK_ERROR(vkGetPhysicalDeviceSurfaceFormatsKHR(gpu.phys_device, surface, &num_surface_formats, gpu.surface_formats.data()), render_engine_initialization_exception);
-
-        uint32_t num_surface_present_modes;
-        NOVA_THROW_IF_VK_ERROR(vkGetPhysicalDeviceSurfacePresentModesKHR(gpu.phys_device, surface, &num_surface_present_modes, nullptr), render_engine_initialization_exception);
-        gpu.present_modes.resize(num_surface_formats);
-        NOVA_THROW_IF_VK_ERROR(vkGetPhysicalDeviceSurfacePresentModesKHR(gpu.phys_device, surface, &num_surface_present_modes, gpu.present_modes.data()), render_engine_initialization_exception);
-
-        swapchain = std::make_unique<swapchain_manager>(3, *this, window->get_window_size());
-    }
-
-    void vulkan_render_engine::open_window(uint32_t width, uint32_t height) {
-#ifdef NOVA_LINUX
-        window = std::make_shared<x11_window>(width, height);
-
-        VkXlibSurfaceCreateInfoKHR x_surface_create_info;
-        x_surface_create_info.sType = VK_STRUCTURE_TYPE_XLIB_SURFACE_CREATE_INFO_KHR;
-        x_surface_create_info.pNext = nullptr;
-        x_surface_create_info.flags = 0;
-        x_surface_create_info.dpy = window->get_display();
-        x_surface_create_info.window = window->get_x11_window();
-
-        NOVA_THROW_IF_VK_ERROR(vkCreateXlibSurfaceKHR(vk_instance, &x_surface_create_info, nullptr, &surface), render_engine_initialization_exception);
-#elif NOVA_WINDOWS
-        window = std::make_shared<win32_window>(width, height);
-
-        VkWin32SurfaceCreateInfoKHR win32_surface_create = {};
-        win32_surface_create.sType = VK_STRUCTURE_TYPE_WIN32_SURFACE_CREATE_INFO_KHR;
-        win32_surface_create.hwnd = window->get_window_handle();
-
-        NOVA_THROW_IF_VK_ERROR(vkCreateWin32SurfaceKHR(vk_instance, &win32_surface_create, nullptr, &surface), render_engine_initialization_exception);
-
-#else
-#error Unsuported window system
-#endif
-    }
-
-    void vulkan_render_engine::validate_mesh_options(const settings_options::block_allocator_settings& options) const {
-        if(options.buffer_part_size % sizeof(full_vertex) != 0) {
-            throw std::runtime_error("vertex_memory_settings.buffer_part_size must be a multiple of sizeof(full_vertex) (which equals " + std::to_string(sizeof(full_vertex)) + ")");
-        }
-        if(options.new_buffer_size % options.buffer_part_size != 0) {
-            throw std::runtime_error("vertex_memory_settings.new_buffer_size must be a multiple of vertex_memory_settings.buffer_part_size (which equals " + std::to_string(options.buffer_part_size) + ")");
-        }
-        if(options.max_total_allocation % options.new_buffer_size != 0) {
-            throw std::runtime_error("vertex_memory_settings.max_total_allocation must be a multiple of vertex_memory_settings.new_buffer_size (which equals " + std::to_string(options.max_total_allocation) + ")");
-        }
-    }
-
-    void vulkan_render_engine::create_device() {
-        uint32_t device_count;
-        NOVA_THROW_IF_VK_ERROR(vkEnumeratePhysicalDevices(vk_instance, &device_count, nullptr), render_engine_initialization_exception);
-        auto* physical_devices = new VkPhysicalDevice[device_count];
-        NOVA_THROW_IF_VK_ERROR(vkEnumeratePhysicalDevices(vk_instance, &device_count, physical_devices), render_engine_initialization_exception);
-
-        uint32_t graphics_family_idx = 0xFFFFFFFF;
-        uint32_t compute_family_idx = 0xFFFFFFFF;
-        uint32_t copy_family_idx = 0xFFFFFFFF;
-
-        for(uint32_t device_idx = 0; device_idx < device_count; device_idx++) {
-            graphics_family_idx = 0xFFFFFFFF;
-            VkPhysicalDevice current_device = physical_devices[device_idx];
-            vkGetPhysicalDeviceProperties(current_device, &gpu.props);
-
-            if(gpu.props.vendorID == 0x8086 && device_count - 1 > device_idx) { // Intel GPU... they are not powerful and we have more available, so skip it
-                continue;
-            }
-
-            if(!does_device_support_extensions(current_device)) {
-                continue;
-            }
-
-            uint32_t queue_family_count;
-            vkGetPhysicalDeviceQueueFamilyProperties(current_device, &queue_family_count, nullptr);
-            gpu.queue_family_props.resize(queue_family_count);
-            vkGetPhysicalDeviceQueueFamilyProperties(current_device, &queue_family_count, gpu.queue_family_props.data());
-
-            for(uint32_t queue_idx = 0; queue_idx < queue_family_count; queue_idx++) {
-                const VkQueueFamilyProperties current_properties = gpu.queue_family_props[queue_idx];
-                if(current_properties.queueCount < 1) {
-                    continue;
-                }
-
-                VkBool32 supports_present = VK_FALSE;
-                NOVA_THROW_IF_VK_ERROR(vkGetPhysicalDeviceSurfaceSupportKHR(current_device, queue_idx, surface, &supports_present), render_engine_initialization_exception);
-                const VkQueueFlags supports_graphics = current_properties.queueFlags & VK_QUEUE_GRAPHICS_BIT;
-                if(supports_graphics && supports_present == VK_TRUE && graphics_family_idx == 0xFFFFFFFF) {
-                    graphics_family_idx = queue_idx;
-                }
-
-                const VkQueueFlags supports_compute = current_properties.queueFlags & VK_QUEUE_COMPUTE_BIT;
-                if(supports_compute && compute_family_idx == 0xFFFFFFFF) {
-                    compute_family_idx = queue_idx;
-                }
-
-                const VkQueueFlags supports_copy = current_properties.queueFlags & VK_QUEUE_TRANSFER_BIT;
-                if(supports_copy && copy_family_idx == 0xFFFFFFFF) {
-                    copy_family_idx = queue_idx;
-                }
-            }
-
-            if(graphics_family_idx != 0xFFFFFFFF) {
-                NOVA_LOG(INFO) << "Selected GPU " << gpu.props.deviceName;
-                gpu.phys_device = current_device;
-                break;
-            }
-        }
-
-        if(!gpu.phys_device) {
-            throw render_engine_initialization_exception("Failed to find good GPU");
-        }
-
-        vkGetPhysicalDeviceFeatures(gpu.phys_device, &gpu.supported_features);
-
-        const float priority = 1.0;
-
-        VkDeviceQueueCreateInfo graphics_queue_create_info{};
-        graphics_queue_create_info.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-        graphics_queue_create_info.pNext = nullptr;
-        graphics_queue_create_info.flags = 0;
-        graphics_queue_create_info.queueCount = 1;
-        graphics_queue_create_info.queueFamilyIndex = graphics_family_idx;
-        graphics_queue_create_info.pQueuePriorities = &priority;
-
-        std::vector<VkDeviceQueueCreateInfo> queue_create_infos = {graphics_queue_create_info};
-
-        VkPhysicalDeviceFeatures physical_device_features{};
-        physical_device_features.geometryShader = VK_TRUE;
-        physical_device_features.tessellationShader = VK_TRUE;
-        physical_device_features.samplerAnisotropy = VK_TRUE;
-
-        VkDeviceCreateInfo device_create_info{};
-        device_create_info.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-        device_create_info.pNext = nullptr;
-        device_create_info.flags = 0;
-        device_create_info.queueCreateInfoCount = static_cast<uint32_t>(queue_create_infos.size());
-        device_create_info.pQueueCreateInfos = queue_create_infos.data();
-        device_create_info.pEnabledFeatures = &physical_device_features;
-        device_create_info.enabledExtensionCount = 1;
-        const char* swapchain_extension = VK_KHR_SWAPCHAIN_EXTENSION_NAME;
-        device_create_info.ppEnabledExtensionNames = &swapchain_extension;
-        device_create_info.enabledLayerCount = static_cast<uint32_t>(enabled_validation_layer_names.size());
-        if(!enabled_validation_layer_names.empty()) {
-            device_create_info.ppEnabledLayerNames = enabled_validation_layer_names.data();
-        }
-
-        NOVA_THROW_IF_VK_ERROR(vkCreateDevice(gpu.phys_device, &device_create_info, nullptr, &device), render_engine_initialization_exception);
-
-        graphics_family_index = graphics_family_idx;
-        vkGetDeviceQueue(device, graphics_family_idx, 0, &graphics_queue);
-        compute_family_index = compute_family_idx;
-        vkGetDeviceQueue(device, compute_family_idx, 0, &compute_queue);
-        copy_family_index = copy_family_idx;
-        vkGetDeviceQueue(device, copy_family_idx, 0, &copy_queue);
-
-        delete[] physical_devices;
-    }
-
     bool vulkan_render_engine::does_device_support_extensions(VkPhysicalDevice device) {
         uint32_t extension_count;
         vkEnumerateDeviceExtensionProperties(device, nullptr, &extension_count, nullptr);
@@ -389,70 +125,6 @@ namespace nova {
         return required.empty();
     }
 
-    void vulkan_render_engine::create_memory_allocator() {
-        VmaAllocatorCreateInfo allocator_create_info = {};
-        allocator_create_info.physicalDevice = gpu.phys_device;
-        allocator_create_info.device = device;
-
-        NOVA_THROW_IF_VK_ERROR(vmaCreateAllocator(&allocator_create_info, &vma_allocator), render_engine_initialization_exception);
-    }
-
-    void vulkan_render_engine::create_global_sync_objects() {
-        VkFenceCreateInfo fence_info = {};
-        fence_info.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-        fence_info.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-
-        VkSemaphoreCreateInfo semaphore_info = {};
-        semaphore_info.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-
-        frame_fences.resize(max_frames_in_queue);
-        image_available_semaphores.resize(max_frames_in_queue);
-        render_finished_semaphores_by_frame.resize(max_frames_in_queue);
-
-        for(uint32_t i = 0; i < max_frames_in_queue; i++) {
-            NOVA_THROW_IF_VK_ERROR(vkCreateFence(device, &fence_info, nullptr, &frame_fences[i]), render_engine_initialization_exception);
-            NOVA_THROW_IF_VK_ERROR(vkCreateSemaphore(device, &semaphore_info, nullptr, &image_available_semaphores[i]), render_engine_initialization_exception);
-        }
-        NOVA_THROW_IF_VK_ERROR(vkCreateFence(device, &fence_info, nullptr, &mesh_rendering_done), render_engine_initialization_exception);
-        NOVA_THROW_IF_VK_ERROR(vkCreateFence(device, &fence_info, nullptr, &upload_to_megamesh_buffer_done), render_engine_initialization_exception);
-    }
-
-    void vulkan_render_engine::set_shaderpack(const shaderpack_data& data) {
-        NOVA_LOG(DEBUG) << "Vulkan render engine loading new shaderpack";
-        if(shaderpack_loaded) {
-            destroy_render_passes();
-            destroy_graphics_pipelines();
-            materials.clear();
-            material_passes_by_pipeline.clear();
-            destroy_dynamic_resources();
-
-            NOVA_LOG(DEBUG) << "Resources from old shaderpacks destroyed";
-        }
-
-        create_textures(data.resources.textures);
-        NOVA_LOG(DEBUG) << "Dynamic textures created";
-        for(const material_data& mat_data : data.materials) {
-            materials[mat_data.name] = mat_data;
-
-            for(const material_pass& mat : mat_data.passes) {
-                material_passes_by_pipeline[mat.pipeline].push_back(mat);
-            }
-        }
-        NOVA_LOG(DEBUG) << "Materials saved";
-
-        create_render_passes(data.passes);
-        NOVA_LOG(DEBUG) << "Created render passes";
-        create_graphics_pipelines(data.pipelines);
-        NOVA_LOG(DEBUG) << "Created pipelines";
-
-        create_material_descriptor_sets();
-        NOVA_LOG(TRACE) << "Material descriptor sets created";
-
-        generate_barriers_for_dynamic_resources();
-
-        shaderpack_loaded = true;
-    }
-
     VkCommandPool vulkan_render_engine::get_command_buffer_pool_for_current_thread(uint32_t queue_index) {
         const std::size_t cur_thread_idx = scheduler->get_current_thread_idx();
         return command_pools_by_thread_idx.at(cur_thread_idx).at(queue_index);
@@ -461,15 +133,6 @@ namespace nova {
     VkDescriptorPool vulkan_render_engine::get_descriptor_pool_for_current_thread() {
         const std::size_t cur_thread_idx = scheduler->get_current_thread_idx();
         return descriptor_pools_by_thread_idx.at(cur_thread_idx);
-    }
-
-    void vulkan_render_engine::create_default_samplers() {
-        VkSamplerCreateInfo point_sampler_create = {};
-        point_sampler_create.sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO;
-        point_sampler_create.magFilter = VK_FILTER_NEAREST;
-        point_sampler_create.minFilter = VK_FILTER_NEAREST;
-
-        NOVA_THROW_IF_VK_ERROR(vkCreateSampler(device, &point_sampler_create, nullptr, &point_sampler), render_engine_initialization_exception);
     }
 
     vk_buffer vulkan_render_engine::get_or_allocate_mesh_staging_buffer(const uint32_t needed_size) {
@@ -555,469 +218,6 @@ namespace nova {
 
     bool vk_resource_binding::operator!=(const vk_resource_binding& other) const {
         return !(*this == other);
-    }
-
-    void vulkan_render_engine::create_render_passes(const std::vector<render_pass_data>& passes) {
-        NOVA_LOG(DEBUG) << "Flattening frame graph...";
-
-        std::unordered_map<std::string, render_pass_data> regular_render_passes;
-        regular_render_passes.reserve(passes.size());
-        render_passes.reserve(passes.size());
-        for(const render_pass_data& pass_data : passes) {
-            render_passes[pass_data.name].data = pass_data;
-            VkFenceCreateInfo fence_info = {};
-            fence_info.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-
-            vkCreateFence(device, &fence_info, nullptr, &render_passes[pass_data.name].fence);
-            regular_render_passes[pass_data.name] = pass_data;
-        }
-
-        render_passes_by_order = order_passes(regular_render_passes);
-
-        const VkExtent2D swapchain_extent = swapchain->get_swapchain_extent();
-
-        for(const auto& [pass_name, pass] : render_passes) {
-            VkSubpassDescription subpass_description;
-            subpass_description.flags = 0;
-            subpass_description.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-            subpass_description.inputAttachmentCount = 0;
-            subpass_description.pInputAttachments = nullptr;
-            subpass_description.preserveAttachmentCount = 0;
-            subpass_description.pPreserveAttachments = nullptr;
-            subpass_description.pResolveAttachments = nullptr;
-            subpass_description.pDepthStencilAttachment = nullptr;
-
-            VkSubpassDependency image_available_dependency;
-            image_available_dependency.dependencyFlags = 0;
-            image_available_dependency.srcSubpass = VK_SUBPASS_EXTERNAL;
-            image_available_dependency.dstSubpass = 0;
-            image_available_dependency.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-            image_available_dependency.srcAccessMask = 0;
-            image_available_dependency.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT;
-            image_available_dependency.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_READ_BIT | VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-
-            VkRenderPassCreateInfo render_pass_create_info;
-            render_pass_create_info.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-            render_pass_create_info.pNext = nullptr;
-            render_pass_create_info.flags = 0;
-            render_pass_create_info.subpassCount = 1;
-            render_pass_create_info.pSubpasses = &subpass_description;
-            render_pass_create_info.dependencyCount = 1;
-            render_pass_create_info.pDependencies = &image_available_dependency;
-
-            std::vector<VkAttachmentReference> attachment_references;
-            std::vector<VkAttachmentDescription> attachments;
-            std::vector<VkImageView> framebuffer_attachments;
-            uint32_t framebuffer_width = 0;
-            uint32_t framebuffer_height = 0;
-
-            std::vector<vk_texture*> textures_in_framebuffer;
-            textures_in_framebuffer.reserve(pass.data.texture_outputs.size() + (pass.data.depth_texture ? 1 : 0));
-            bool writes_to_backbuffer = false;
-            // Collect framebuffer size information from color output attachments
-            for(const texture_attachment& attachment : pass.data.texture_outputs) {
-                if(attachment.name == "Backbuffer") {
-                    // Handle backbuffer
-                    // Backbuffer framebuffers are handled by themselves in their own special snowflake way so we just need to skip everything
-                    writes_to_backbuffer = true;
-
-                    VkAttachmentDescription desc = {};
-                    desc.flags = 0;
-                    desc.format = swapchain->get_swapchain_format();
-                    desc.samples = VK_SAMPLE_COUNT_1_BIT;
-                    desc.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-                    desc.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-                    desc.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-                    desc.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-                    desc.initialLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-                    desc.finalLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-                    attachments.push_back(desc);
-
-                    VkAttachmentReference ref = {};
-
-                    ref.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-                    ref.attachment = static_cast<uint32_t>(attachments.size()) - 1;
-
-                    attachment_references.push_back(ref);
-
-                    framebuffer_width = swapchain_extent.width;
-                    framebuffer_height = swapchain_extent.height;
-
-                    render_passes[pass_name].writes_to_backbuffer = true;
-
-                    break;
-
-                } else {
-                    const vk_texture& tex = textures.at(attachment.name);
-                    // the textures array _should_ _never_ change from this point onward. If it does, this pointer will probably point to undefined data
-                    textures_in_framebuffer.push_back(&textures.at(attachment.name));
-
-                    framebuffer_attachments.push_back(tex.image_view);
-
-                    const glm::uvec2 attachment_size = tex.data.format.get_size_in_pixels({swapchain_extent.width, swapchain_extent.height});
-
-                    if(framebuffer_width == 0) {
-                        framebuffer_width = attachment_size.x;
-
-                    } else if(attachment_size.x != framebuffer_width) {
-                        NOVA_LOG(ERROR) << "Texture " << attachment.name << " used by renderpass " << pass_name << " has a width of " << attachment_size.x << ", but the framebuffer has a width of " << framebuffer_width
-                                        << ". This is illegal, all input textures of a single renderpass must be the same size";
-                    }
-
-                    if(framebuffer_height == 0) {
-                        framebuffer_height = attachment_size.y;
-
-                    } else if(attachment_size.y != framebuffer_height) {
-                        NOVA_LOG(ERROR) << "Texture " << attachment.name << " used by renderpass " << pass_name << " has a height of " << attachment_size.y << ", but the framebuffer has a height of " << framebuffer_height
-                                        << ". This is illegal, all input textures of a single renderpass must be the same size";
-                    }
-
-                    NOVA_LOG(TRACE) << "Adding image view " << textures.at(attachment.name).image_view << " from image " << attachment.name << " to framebuffer for renderpass " << pass.data.name;
-
-                    VkAttachmentDescription desc = {};
-                    desc.flags = 0;
-                    desc.format = tex.format;
-                    desc.samples = VK_SAMPLE_COUNT_1_BIT;
-                    desc.loadOp = attachment.clear ? VK_ATTACHMENT_LOAD_OP_CLEAR : VK_ATTACHMENT_LOAD_OP_LOAD;
-                    desc.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-                    desc.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-                    desc.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-                    desc.initialLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-                    desc.finalLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-                    attachments.push_back(desc);
-
-                    VkAttachmentReference ref = {};
-
-                    ref.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-                    ref.attachment = static_cast<uint32_t>(attachments.size()) - 1;
-
-                    attachment_references.push_back(ref);
-                }
-            }
-
-            VkAttachmentReference depth_reference = {};
-            // Collect framebuffer size information from the depth attachment
-            if(pass.data.depth_texture) {
-
-                const vk_texture& tex = textures.at(pass.data.depth_texture->name);
-                framebuffer_attachments.push_back(tex.image_view);
-
-                textures_in_framebuffer.push_back(&textures.at(pass.data.depth_texture->name));
-
-                const glm::uvec2 attachment_size = tex.data.format.get_size_in_pixels({swapchain_extent.width, swapchain_extent.height});
-
-                if(framebuffer_width == 0) {
-                    framebuffer_width = attachment_size.x;
-
-                } else if(attachment_size.x != framebuffer_width) {
-                    NOVA_LOG(ERROR) << "Texture " << pass.data.depth_texture->name << " used by renderpass " << pass_name << " has a width of " << attachment_size.x << ", but the framebuffer has a width of " << framebuffer_width
-                                    << ". This is illegal, all input textures of a single renderpass must be the same size";
-                }
-
-                if(framebuffer_height == 0) {
-                    framebuffer_height = attachment_size.y;
-
-                } else if(attachment_size.y != framebuffer_height) {
-                    NOVA_LOG(ERROR) << "Texture " << pass.data.depth_texture->name << " used by renderpass " << pass_name << " has a height of " << attachment_size.y << ", but the framebuffer has a height of " << framebuffer_height
-                                    << ". This is illegal, all input textures of a single renderpass must be the same size";
-                }
-
-                VkAttachmentDescription desc = {};
-                desc.flags = 0;
-                desc.format = tex.format;
-                desc.samples = VK_SAMPLE_COUNT_1_BIT;
-                desc.loadOp = pass.data.depth_texture->clear ? VK_ATTACHMENT_LOAD_OP_CLEAR : VK_ATTACHMENT_LOAD_OP_LOAD;
-                desc.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-                desc.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-                desc.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-                desc.initialLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-                desc.finalLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-
-                attachments.push_back(desc);
-
-                depth_reference.layout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-                depth_reference.attachment = static_cast<uint32_t>(attachments.size()) - 1;
-
-                subpass_description.pDepthStencilAttachment = &depth_reference;
-            }
-
-            if(framebuffer_width == 0) {
-                NOVA_LOG(ERROR) << "Framebuffer width for pass " << pass.data.name
-                                << " is 0. This is illegal! Make sure that there is at least one attachment for this render pass, and ensure that all attachments used by this pass have a non-zero width";
-            }
-
-            if(framebuffer_height == 0) {
-                NOVA_LOG(ERROR) << "Framebuffer height for pass " << pass.data.name
-                                << " is 0. This is illegal! Make sure that there is at least one attachment for this render pass, and ensure that all attachments used by this pass have a non-zero height";
-            }
-
-            if(framebuffer_attachments.size() > gpu.props.limits.maxColorAttachments) {
-                NOVA_LOG(ERROR) << "Framebuffer for pass " << pass.data.name << " has " << framebuffer_attachments.size() << " color attachments, but your GPU only supports " << gpu.props.limits.maxColorAttachments
-                                << ". Please reduce the number of attachments that this pass uses, possibly by changing some of your input attachments to bound textures";
-            }
-
-            subpass_description.colorAttachmentCount = static_cast<uint32_t>(attachment_references.size());
-            subpass_description.pColorAttachments = attachment_references.data();
-
-            render_pass_create_info.attachmentCount = static_cast<uint32_t>(attachments.size());
-            render_pass_create_info.pAttachments = attachments.data();
-
-            NOVA_THROW_IF_VK_ERROR(vkCreateRenderPass(device, &render_pass_create_info, nullptr, &render_passes[pass_name].pass), render_engine_initialization_exception);
-
-            if(writes_to_backbuffer) {
-                if(pass.data.texture_outputs.size() > 1) {
-                    NOVA_LOG(ERROR) << "Pass " << pass.data.name << " writes to the backbuffer, and other textures. Passes that write to the backbuffer are not allowed to write to any other textures";
-                }
-
-                render_passes[pass_name].framebuffer.framebuffer = nullptr;
-
-            } else {
-                VkFramebufferCreateInfo framebuffer_create_info = {};
-                framebuffer_create_info.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
-                framebuffer_create_info.renderPass = render_passes[pass_name].pass;
-                framebuffer_create_info.attachmentCount = static_cast<uint32_t>(framebuffer_attachments.size());
-                framebuffer_create_info.pAttachments = framebuffer_attachments.data();
-                framebuffer_create_info.width = framebuffer_width;
-                framebuffer_create_info.height = framebuffer_height;
-                framebuffer_create_info.layers = 1;
-
-                std::stringstream ss;
-                for(const VkImageView& attachment : framebuffer_attachments) {
-                    ss << attachment << ", ";
-                }
-
-                NOVA_LOG(TRACE) << "Creating framebuffer with size (" << framebuffer_width << ", " << framebuffer_height << "), and attachments " << ss.str();
-
-                VkFramebuffer framebuffer;
-                NOVA_THROW_IF_VK_ERROR(vkCreateFramebuffer(device, &framebuffer_create_info, nullptr, &framebuffer), render_engine_initialization_exception);
-                render_passes[pass_name].framebuffer.framebuffer = framebuffer;
-                render_passes[pass_name].framebuffer.images = std::move(textures_in_framebuffer);
-            }
-
-            render_passes[pass_name].render_area = {{0, 0}, {framebuffer_width, framebuffer_height}};
-
-            VkDebugUtilsObjectNameInfoEXT object_name = {};
-            object_name.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT;
-            object_name.objectType = VK_OBJECT_TYPE_IMAGE;
-            object_name.objectHandle = reinterpret_cast<uint64_t>(render_passes[pass_name].pass);
-            object_name.pObjectName = pass_name.c_str();
-            NOVA_THROW_IF_VK_ERROR(vkSetDebugUtilsObjectNameEXT(device, &object_name), render_engine_initialization_exception);
-        }
-    }
-
-    void vulkan_render_engine::create_graphics_pipelines(const std::vector<pipeline_data>& pipelines) {
-        const VkExtent2D swapchain_extent = swapchain->get_swapchain_extent();
-
-        for(const pipeline_data& data : pipelines) {
-            NOVA_LOG(TRACE) << "Creating a VkPipeline for pipeline " << data.name;
-            vk_pipeline nova_pipeline;
-            nova_pipeline.data = data;
-
-            std::vector<VkPipelineShaderStageCreateInfo> shader_stages;
-            std::unordered_map<VkShaderStageFlags, VkShaderModule> shader_modules;
-
-            NOVA_LOG(TRACE) << "Compiling vertex module";
-            shader_modules[VK_SHADER_STAGE_VERTEX_BIT] = create_shader_module(data.vertex_shader.source);
-            get_shader_module_descriptors(data.vertex_shader.source, nova_pipeline.bindings);
-
-            if(data.geometry_shader) {
-                NOVA_LOG(TRACE) << "Compiling geometry module";
-                shader_modules[VK_SHADER_STAGE_GEOMETRY_BIT] = create_shader_module(data.geometry_shader->source);
-                get_shader_module_descriptors(data.geometry_shader->source, nova_pipeline.bindings);
-            }
-
-            if(data.tessellation_control_shader) {
-                NOVA_LOG(TRACE) << "Compiling tessellation_control module";
-                shader_modules[VK_SHADER_STAGE_TESSELLATION_CONTROL_BIT] = create_shader_module(data.tessellation_control_shader->source);
-                get_shader_module_descriptors(data.tessellation_control_shader->source, nova_pipeline.bindings);
-            }
-
-            if(data.tessellation_evaluation_shader) {
-                NOVA_LOG(TRACE) << "Compiling tessellation_evaluation module";
-                shader_modules[VK_SHADER_STAGE_TESSELLATION_EVALUATION_BIT] = create_shader_module(data.tessellation_evaluation_shader->source);
-                get_shader_module_descriptors(data.tessellation_evaluation_shader->source, nova_pipeline.bindings);
-            }
-
-            if(data.fragment_shader) {
-                NOVA_LOG(TRACE) << "Compiling fragment module";
-                shader_modules[VK_SHADER_STAGE_FRAGMENT_BIT] = create_shader_module(data.fragment_shader->source);
-                get_shader_module_descriptors(data.fragment_shader->source, nova_pipeline.bindings);
-            }
-
-            nova_pipeline.layouts = create_descriptor_set_layouts(nova_pipeline.bindings);
-
-            VkPipelineLayoutCreateInfo pipeline_layout_create_info;
-            pipeline_layout_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-            pipeline_layout_create_info.pNext = nullptr;
-            pipeline_layout_create_info.flags = 0;
-            pipeline_layout_create_info.setLayoutCount = static_cast<uint32_t>(nova_pipeline.layouts.size());
-            pipeline_layout_create_info.pSetLayouts = nova_pipeline.layouts.data();
-            pipeline_layout_create_info.pushConstantRangeCount = 0;
-            pipeline_layout_create_info.pPushConstantRanges = nullptr;
-
-            VkPipelineLayout layout;
-            NOVA_THROW_IF_VK_ERROR(vkCreatePipelineLayout(device, &pipeline_layout_create_info, nullptr, &layout), render_engine_initialization_exception);
-            nova_pipeline.layout = layout;
-
-            for(const auto& [stage, shader_module] : shader_modules) {
-                VkPipelineShaderStageCreateInfo shader_stage_create_info;
-                shader_stage_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-                shader_stage_create_info.pNext = nullptr;
-                shader_stage_create_info.flags = 0;
-                shader_stage_create_info.stage = static_cast<VkShaderStageFlagBits>(stage);
-                shader_stage_create_info.module = shader_module;
-                shader_stage_create_info.pName = "main";
-                shader_stage_create_info.pSpecializationInfo = nullptr;
-
-                shader_stages.push_back(shader_stage_create_info);
-            }
-
-            const std::vector<VkVertexInputBindingDescription>& vertex_binding_descriptions = vulkan::get_vertex_input_binding_descriptions();
-            const std::vector<VkVertexInputAttributeDescription>& vertex_attribute_descriptions = vulkan::get_vertex_input_attribute_descriptions();
-
-            VkPipelineVertexInputStateCreateInfo vertex_input_state_create_info;
-            vertex_input_state_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-            vertex_input_state_create_info.pNext = nullptr;
-            vertex_input_state_create_info.flags = 0;
-            vertex_input_state_create_info.vertexBindingDescriptionCount = static_cast<uint32_t>(vertex_binding_descriptions.size());
-            vertex_input_state_create_info.pVertexBindingDescriptions = vertex_binding_descriptions.data();
-            vertex_input_state_create_info.vertexAttributeDescriptionCount = static_cast<uint32_t>(vertex_attribute_descriptions.size());
-            vertex_input_state_create_info.pVertexAttributeDescriptions = vertex_attribute_descriptions.data();
-
-            VkPipelineInputAssemblyStateCreateInfo input_assembly_create_info;
-            input_assembly_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-            input_assembly_create_info.pNext = nullptr;
-            input_assembly_create_info.flags = 0;
-            input_assembly_create_info.primitiveRestartEnable = VK_FALSE;
-            switch(data.primitive_mode) {
-                case primitive_topology_enum::Triangles:
-                    input_assembly_create_info.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-                    break;
-                case primitive_topology_enum::Lines:
-                    input_assembly_create_info.topology = VK_PRIMITIVE_TOPOLOGY_LINE_LIST;
-                    break;
-            }
-
-            VkViewport viewport;
-            viewport.x = 0;
-            viewport.y = 0;
-            viewport.width = static_cast<float>(swapchain_extent.width);
-            viewport.height = static_cast<float>(swapchain_extent.height);
-            viewport.minDepth = 0.0f;
-            viewport.maxDepth = 1.0f;
-
-            VkRect2D scissor;
-            scissor.offset = {0, 0};
-            scissor.extent = swapchain_extent;
-
-            VkPipelineViewportStateCreateInfo viewport_state_create_info;
-            viewport_state_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-            viewport_state_create_info.pNext = nullptr;
-            viewport_state_create_info.flags = 0;
-            viewport_state_create_info.viewportCount = 1;
-            viewport_state_create_info.pViewports = &viewport;
-            viewport_state_create_info.scissorCount = 1;
-            viewport_state_create_info.pScissors = &scissor;
-
-            VkPipelineRasterizationStateCreateInfo rasterizer_create_info;
-            rasterizer_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-            rasterizer_create_info.pNext = nullptr;
-            rasterizer_create_info.flags = 0;
-            rasterizer_create_info.depthClampEnable = VK_FALSE;
-            rasterizer_create_info.rasterizerDiscardEnable = VK_FALSE;
-            rasterizer_create_info.polygonMode = VK_POLYGON_MODE_FILL;
-            rasterizer_create_info.lineWidth = 1.0f;
-            rasterizer_create_info.cullMode = VK_CULL_MODE_BACK_BIT;
-            rasterizer_create_info.frontFace = VK_FRONT_FACE_CLOCKWISE;
-            rasterizer_create_info.depthBiasEnable = VK_TRUE;
-            rasterizer_create_info.depthClampEnable = VK_FALSE;
-            rasterizer_create_info.depthBiasConstantFactor = data.depth_bias;
-            rasterizer_create_info.depthBiasClamp = 0.0f;
-            rasterizer_create_info.depthBiasSlopeFactor = data.slope_scaled_depth_bias;
-
-            VkPipelineMultisampleStateCreateInfo multisample_create_info;
-            multisample_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-            multisample_create_info.pNext = nullptr;
-            multisample_create_info.flags = 0;
-            multisample_create_info.sampleShadingEnable = VK_FALSE;
-            multisample_create_info.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-            multisample_create_info.minSampleShading = 1.0f;
-            multisample_create_info.pSampleMask = nullptr;
-            multisample_create_info.alphaToCoverageEnable = VK_FALSE;
-            multisample_create_info.alphaToOneEnable = VK_FALSE;
-
-            VkPipelineDepthStencilStateCreateInfo depth_stencil_create_info = {};
-            depth_stencil_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO;
-            depth_stencil_create_info.depthTestEnable = std::find(data.states.begin(), data.states.end(), state_enum::DisableDepthTest) == data.states.end();
-            depth_stencil_create_info.depthWriteEnable = std::find(data.states.begin(), data.states.end(), state_enum::DisableDepthWrite) == data.states.end();
-            depth_stencil_create_info.depthCompareOp = vulkan::type_converters::to_compare_op(data.depth_func);
-            depth_stencil_create_info.depthBoundsTestEnable = VK_FALSE;
-            depth_stencil_create_info.stencilTestEnable = std::find(data.states.begin(), data.states.end(), state_enum::EnableStencilTest) != data.states.end();
-            if(data.front_face) {
-                depth_stencil_create_info.front.failOp = vulkan::type_converters::to_stencil_op(data.front_face->fail_op);
-                depth_stencil_create_info.front.passOp = vulkan::type_converters::to_stencil_op(data.front_face->pass_op);
-                depth_stencil_create_info.front.depthFailOp = vulkan::type_converters::to_stencil_op(data.front_face->depth_fail_op);
-                depth_stencil_create_info.front.compareOp = vulkan::type_converters::to_compare_op(data.front_face->compare_op);
-                depth_stencil_create_info.front.compareMask = data.front_face->compare_mask;
-                depth_stencil_create_info.front.writeMask = data.front_face->write_mask;
-            }
-            if(data.back_face) {
-                depth_stencil_create_info.back.failOp = vulkan::type_converters::to_stencil_op(data.back_face->fail_op);
-                depth_stencil_create_info.back.passOp = vulkan::type_converters::to_stencil_op(data.back_face->pass_op);
-                depth_stencil_create_info.back.depthFailOp = vulkan::type_converters::to_stencil_op(data.back_face->depth_fail_op);
-                depth_stencil_create_info.back.compareOp = vulkan::type_converters::to_compare_op(data.back_face->compare_op);
-                depth_stencil_create_info.back.compareMask = data.back_face->compare_mask;
-                depth_stencil_create_info.back.writeMask = data.back_face->write_mask;
-            }
-
-            VkPipelineColorBlendAttachmentState color_blend_attachment;
-            color_blend_attachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-            color_blend_attachment.blendEnable = VK_TRUE;
-            color_blend_attachment.srcColorBlendFactor = vulkan::type_converters::to_blend_factor(data.source_blend_factor);
-            color_blend_attachment.dstColorBlendFactor = vulkan::type_converters::to_blend_factor(data.destination_blend_factor);
-            color_blend_attachment.colorBlendOp = VK_BLEND_OP_ADD;
-            color_blend_attachment.srcAlphaBlendFactor = vulkan::type_converters::to_blend_factor(data.alpha_src);
-            color_blend_attachment.dstAlphaBlendFactor = vulkan::type_converters::to_blend_factor(data.alpha_dst);
-            color_blend_attachment.alphaBlendOp = VK_BLEND_OP_ADD;
-
-            VkPipelineColorBlendStateCreateInfo color_blend_create_info;
-            color_blend_create_info.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-            color_blend_create_info.pNext = nullptr;
-            color_blend_create_info.flags = 0;
-            color_blend_create_info.logicOpEnable = VK_FALSE;
-            color_blend_create_info.logicOp = VK_LOGIC_OP_COPY;
-            color_blend_create_info.attachmentCount = 1;
-            color_blend_create_info.pAttachments = &color_blend_attachment;
-            color_blend_create_info.blendConstants[0] = 0.0f;
-            color_blend_create_info.blendConstants[1] = 0.0f;
-            color_blend_create_info.blendConstants[2] = 0.0f;
-            color_blend_create_info.blendConstants[3] = 0.0f;
-
-            VkGraphicsPipelineCreateInfo pipeline_create_info = {};
-            pipeline_create_info.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-            pipeline_create_info.pNext = nullptr;
-            pipeline_create_info.flags = 0;
-            pipeline_create_info.stageCount = static_cast<uint32_t>(shader_stages.size());
-            pipeline_create_info.pStages = shader_stages.data();
-            pipeline_create_info.pVertexInputState = &vertex_input_state_create_info;
-            pipeline_create_info.pInputAssemblyState = &input_assembly_create_info;
-            pipeline_create_info.pViewportState = &viewport_state_create_info;
-            pipeline_create_info.pRasterizationState = &rasterizer_create_info;
-            pipeline_create_info.pMultisampleState = &multisample_create_info;
-            pipeline_create_info.pDepthStencilState = &depth_stencil_create_info;
-            pipeline_create_info.pColorBlendState = &color_blend_create_info;
-            pipeline_create_info.pDynamicState = nullptr;
-            pipeline_create_info.layout = layout;
-            pipeline_create_info.renderPass = render_passes.at(data.pass).pass;
-            pipeline_create_info.subpass = 0;
-            pipeline_create_info.basePipelineIndex = -1;
-
-            NOVA_THROW_IF_VK_ERROR(vkCreateGraphicsPipelines(device, VK_NULL_HANDLE, 1, &pipeline_create_info, nullptr, &nova_pipeline.pipeline), render_engine_initialization_exception);
-
-            pipelines_by_renderpass[data.pass].push_back(nova_pipeline);
-        }
     }
 
     VkShaderModule vulkan_render_engine::create_shader_module(const std::vector<uint32_t>& spirv) const {
@@ -1261,15 +461,6 @@ namespace nova {
         }
 
         return {attachment_descriptions, attachment_references};
-    }
-
-    void vulkan_render_engine::destroy_render_passes() {
-        for(const auto& [pass_name, pass] : render_passes) {
-            vkDestroyRenderPass(device, pass.pass, nullptr);
-        }
-
-        render_passes.clear();
-        render_passes_by_order.clear();
     }
 
     void vulkan_render_engine::destroy_graphics_pipelines() {
@@ -1586,142 +777,6 @@ namespace nova {
         render_finished_semaphores_by_frame[current_frame].clear();
     }
 
-    void vulkan_render_engine::create_per_thread_command_pools() {
-        const uint32_t num_threads = scheduler->get_num_threads();
-        command_pools_by_thread_idx.reserve(num_threads);
-
-        for(uint32_t i = 0; i < num_threads; i++) {
-            command_pools_by_thread_idx.push_back(make_new_command_pools());
-        }
-    }
-
-    void vulkan_render_engine::create_per_thread_descriptor_pools() {
-        const uint32_t num_threads = scheduler->get_num_threads();
-        descriptor_pools_by_thread_idx.reserve(num_threads);
-
-        for(uint32_t i = 0; i < num_threads; i++) {
-            descriptor_pools_by_thread_idx.push_back(make_new_descriptor_pool());
-        }
-    }
-
-    std::unordered_map<uint32_t, VkCommandPool> vulkan_render_engine::make_new_command_pools() const {
-        std::vector<uint32_t> queue_indices;
-        queue_indices.push_back(graphics_family_index);
-        queue_indices.push_back(copy_family_index);
-        queue_indices.push_back(compute_family_index);
-
-        std::unordered_map<uint32_t, VkCommandPool> pools_by_queue;
-        pools_by_queue.reserve(3);
-
-        for(const uint32_t queue_index : queue_indices) {
-            VkCommandPoolCreateInfo command_pool_create_info;
-            command_pool_create_info.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-            command_pool_create_info.pNext = nullptr;
-            command_pool_create_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-            command_pool_create_info.queueFamilyIndex = queue_index;
-
-            VkCommandPool command_pool;
-            NOVA_THROW_IF_VK_ERROR(vkCreateCommandPool(device, &command_pool_create_info, nullptr, &command_pool), render_engine_initialization_exception);
-            pools_by_queue[queue_index] = command_pool;
-        }
-
-        return pools_by_queue;
-    }
-
-    VkDescriptorPool vulkan_render_engine::make_new_descriptor_pool() const {
-        std::vector<VkDescriptorPoolSize> pool_sizes;
-        pool_sizes.emplace_back(VkDescriptorPoolSize{VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, 5}); // Virtual textures greatly reduces the number of total textures
-        pool_sizes.emplace_back(VkDescriptorPoolSize{VK_DESCRIPTOR_TYPE_SAMPLER, 5});
-        pool_sizes.emplace_back(VkDescriptorPoolSize{VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 5000});
-
-        VkDescriptorPoolCreateInfo pool_create_info = {};
-        pool_create_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-        pool_create_info.maxSets = 5000;
-        pool_create_info.poolSizeCount = static_cast<uint32_t>(pool_sizes.size());
-        pool_create_info.pPoolSizes = pool_sizes.data();
-
-        VkDescriptorPool pool;
-        NOVA_THROW_IF_VK_ERROR(vkCreateDescriptorPool(device, &pool_create_info, nullptr, &pool), descriptor_pool_creation_failed);
-
-        return pool;
-    }
-
-    void vulkan_render_engine::create_textures(const std::vector<texture_resource_data>& texture_datas) {
-        const VkExtent2D swapchain_extent = swapchain->get_swapchain_extent();
-        const glm::uvec2 swapchain_extent_glm = {swapchain_extent.width, swapchain_extent.height};
-        for(const texture_resource_data& texture_data : texture_datas) {
-            vk_texture texture;
-            texture.is_dynamic = true;
-            texture.data = texture_data;
-            texture.format = to_vk_format(texture_data.format.pixel_format);
-            texture.layout = VK_IMAGE_LAYOUT_UNDEFINED;
-
-            VkImageCreateInfo image_create_info = {};
-            image_create_info.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
-            image_create_info.imageType = VK_IMAGE_TYPE_2D;
-            image_create_info.format = texture.format;
-            const glm::uvec2 texture_size = texture_data.format.get_size_in_pixels(swapchain_extent_glm);
-            image_create_info.extent.width = texture_size.x;
-            image_create_info.extent.height = texture_size.y;
-            image_create_info.extent.depth = 1;
-            image_create_info.mipLevels = 1;
-            image_create_info.arrayLayers = 1;
-            image_create_info.samples = VK_SAMPLE_COUNT_1_BIT;
-            image_create_info.usage = VK_IMAGE_USAGE_SAMPLED_BIT;
-            if(texture.format == VK_FORMAT_D24_UNORM_S8_UINT || texture.format == VK_FORMAT_D32_SFLOAT) {
-                image_create_info.usage |= VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT;
-
-            } else {
-                image_create_info.usage |= VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-            }
-            image_create_info.queueFamilyIndexCount = 1;
-            image_create_info.pQueueFamilyIndices = &graphics_family_index;
-            image_create_info.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-
-            VmaAllocationCreateInfo alloc_create_info = {};
-            alloc_create_info.flags = VMA_ALLOCATION_CREATE_DEDICATED_MEMORY_BIT;
-            alloc_create_info.usage = VMA_MEMORY_USAGE_GPU_ONLY;
-            alloc_create_info.requiredFlags = VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT;
-            alloc_create_info.preferredFlags = 0;
-            alloc_create_info.memoryTypeBits = 0;
-            alloc_create_info.pool = VK_NULL_HANDLE;
-            alloc_create_info.pUserData = nullptr;
-
-            vmaCreateImage(vma_allocator, &image_create_info, &alloc_create_info, &texture.image, &texture.allocation, &texture.vma_info);
-
-            VkImageViewCreateInfo image_view_create_info = {};
-            image_view_create_info.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-            image_view_create_info.image = texture.image;
-            image_view_create_info.viewType = VK_IMAGE_VIEW_TYPE_2D;
-            image_view_create_info.format = image_create_info.format;
-            if(texture.format == VK_FORMAT_D24_UNORM_S8_UINT || texture.format == VK_FORMAT_D32_SFLOAT) {
-                image_view_create_info.subresourceRange.aspectMask = VK_IMAGE_ASPECT_DEPTH_BIT;
-                texture.is_depth_tex = true;
-
-            } else {
-                image_view_create_info.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-            }
-            image_view_create_info.subresourceRange.baseArrayLayer = 0;
-            image_view_create_info.subresourceRange.layerCount = 1;
-            image_view_create_info.subresourceRange.baseMipLevel = 0;
-            image_view_create_info.subresourceRange.levelCount = 1;
-
-            vkCreateImageView(device, &image_view_create_info, nullptr, &texture.image_view);
-
-            VkDebugUtilsObjectNameInfoEXT object_name = {};
-            object_name.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT;
-            object_name.objectType = VK_OBJECT_TYPE_IMAGE;
-            object_name.objectHandle = reinterpret_cast<uint64_t>(texture.image);
-            object_name.pObjectName = texture_data.name.c_str();
-            NOVA_THROW_IF_VK_ERROR(vkSetDebugUtilsObjectNameEXT(device, &object_name), render_engine_initialization_exception);
-            NOVA_LOG(INFO) << "Set object " << texture.image << " to have name " << texture_data.name;
-
-            textures[texture_data.name] = texture;
-        }
-
-        dynamic_textures_need_to_transition = true;
-    }
-
     void vulkan_render_engine::add_resource_to_bindings(std::unordered_map<std::string, vk_resource_binding>& bindings, const spirv_cross::CompilerGLSL& shader_compiler, const spirv_cross::Resource& resource, const VkDescriptorType type) {
         const uint32_t set = shader_compiler.get_decoration(resource.id, spv::DecorationDescriptorSet);
         const uint32_t binding = shader_compiler.get_decoration(resource.id, spv::DecorationBinding);
@@ -1803,43 +858,6 @@ namespace nova {
         return layouts;
     }
 
-    void vulkan_render_engine::create_material_descriptor_sets() {
-        for(const auto& [renderpass_name, pipelines] : pipelines_by_renderpass) {
-            for(const auto& pipeline : pipelines) {
-                std::vector<material_pass>& material_passes = material_passes_by_pipeline.at(pipeline.data.name);
-                for(material_pass& mat_pass : material_passes) {
-                    if(pipeline.layouts.empty()) {
-                        // If there's no layouts, we're done
-                        NOVA_LOG(TRACE) << "No layouts for pipeline " << pipeline.data.name << ", which material pass " << mat_pass.name << " of material " << mat_pass.material_name << " uses";
-                        continue;
-                    }
-
-                    NOVA_LOG(TRACE) << "Creating descriptor sets for pipeline " << pipeline.data.name;
-
-                    auto layouts = std::vector<VkDescriptorSetLayout>{};
-                    layouts.reserve(pipeline.layouts.size());
-
-                    // CLion might tell you to simplify this into a foreach loop... DO NOT! The layouts need to be added in set
-                    // order, not map order which is what you'll get if you use a foreach - AND IT'S WRONG
-                    for(uint32_t i = 0; i < pipeline.layouts.size(); i++) {
-                        layouts.push_back(pipeline.layouts.at(static_cast<int32_t>(i)));
-                    }
-
-                    VkDescriptorSetAllocateInfo alloc_info = {};
-                    alloc_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-                    alloc_info.descriptorPool = get_descriptor_pool_for_current_thread();
-                    alloc_info.descriptorSetCount = static_cast<uint32_t>(layouts.size());
-                    alloc_info.pSetLayouts = layouts.data();
-
-                    mat_pass.descriptor_sets.resize(layouts.size());
-                    NOVA_THROW_IF_VK_ERROR(vkAllocateDescriptorSets(device, &alloc_info, mat_pass.descriptor_sets.data()), shaderpack_loading_error);
-
-                    update_material_descriptor_sets(mat_pass, pipeline.bindings);
-                }
-            }
-        }
-    }
-
     std::vector<VkImageMemoryBarrier> vulkan_render_engine::make_attachment_to_shader_read_only_barriers(const std::unordered_set<std::string>& textures) {
         std::vector<VkImageMemoryBarrier> barriers;
         barriers.reserve(textures.size());
@@ -1873,208 +891,6 @@ namespace nova {
         }
 
         return barriers;
-    }
-
-    enum class barrier_necessity { maybe, yes, no };
-
-    void vulkan_render_engine::create_barriers_for_renderpass(vk_render_pass& pass) {
-        /*
-         * For each renderpass:
-         * - Walk backwards through previous renderpasses
-         * - If we find one that writes to a resource we read from before we find one that reads from a resource we read from, add a barrier
-         * - If we find one that reads from a resource we write to before we find one that writes to a resource we write to, add a barrier
-         */
-
-        std::unordered_map<std::string, barrier_necessity> read_texture_barrier_necessity;
-        read_texture_barrier_necessity.reserve(pass.data.texture_inputs.size());
-        for(const std::string& tex_name : pass.data.texture_inputs) {
-            read_texture_barrier_necessity[tex_name] = barrier_necessity::maybe;
-        }
-
-        std::unordered_map<std::string, barrier_necessity> write_texture_barrier_necessity;
-        write_texture_barrier_necessity.reserve(pass.data.texture_outputs.size());
-        for(const texture_attachment& attach : pass.data.texture_outputs) {
-            write_texture_barrier_necessity[attach.name] = barrier_necessity::maybe;
-        }
-
-        // Find where we are in the list of passes
-        uint64_t idx = 0;
-        for(; idx < render_passes_by_order.size(); idx++) {
-            if(render_passes_by_order.at(idx) == pass.data.name) {
-                break;
-            }
-        }
-
-        // Walk backwards from where we are, checking if any of the textures the previous pass writes to are textures we need to barrier
-        while(true) {
-            const std::string& prev_pass_name = render_passes_by_order.at(idx);
-            const vk_render_pass& previous_pass = render_passes.at(prev_pass_name);
-            // If the previous pass reads from a texture that we read from, it (or a even earlier pass) will have the barrier
-            const std::vector<std::string> read_textures = previous_pass.data.texture_inputs;
-
-            for(const std::string& prev_read_texture : read_textures) {
-                // If the previous pass reads from a texture that we read from, we don't need to barrier it
-                if(read_texture_barrier_necessity.find(prev_read_texture) != read_texture_barrier_necessity.end()) {
-                    if(read_texture_barrier_necessity.at(prev_read_texture) == barrier_necessity::maybe) {
-                        NOVA_LOG(TRACE) << "Do not need a barrier for read texture " << prev_read_texture << " before renderpass " << pass.data.name << " because pass " << previous_pass.data.name << " reads from it as well";
-                        read_texture_barrier_necessity[prev_read_texture] = barrier_necessity::no;
-                    }
-                }
-
-                // If the previous pass reads from a texture that we write to, we need a barrier
-                if(write_texture_barrier_necessity.find(prev_read_texture) != write_texture_barrier_necessity.end()) {
-                    if(write_texture_barrier_necessity.at(prev_read_texture) == barrier_necessity::maybe) {
-                        NOVA_LOG(TRACE) << "Need a barrier for write texture " << prev_read_texture << " before renderpass " << pass.data.name << " because pass " << previous_pass.data.name << " reads from it";
-                        write_texture_barrier_necessity[prev_read_texture] = barrier_necessity::no;
-                    }
-                }
-            }
-
-            for(const texture_attachment& prev_write_tex : previous_pass.data.texture_outputs) {
-                // If the previous pass write to a texture that we read from, we need a barrier
-                if(read_texture_barrier_necessity.find(prev_write_tex.name) != read_texture_barrier_necessity.end()) {
-                    if(read_texture_barrier_necessity.at(prev_write_tex.name) == barrier_necessity::maybe) {
-                        NOVA_LOG(TRACE) << "Need a barrier for read texture " << prev_write_tex.name << " before renderpass " << pass.data.name << " because pass " << previous_pass.data.name << " writes to is";
-                        read_texture_barrier_necessity[prev_write_tex.name] = barrier_necessity::yes;
-                    }
-                }
-
-                // If the previous pass write to a texture that we write to, we don't need to barrier it
-                if(write_texture_barrier_necessity.find(prev_write_tex.name) != write_texture_barrier_necessity.end()) {
-                    if(write_texture_barrier_necessity.at(prev_write_tex.name) == barrier_necessity::maybe) {
-                        NOVA_LOG(TRACE) << "Do not need a barrier for write texture " << prev_write_tex.name << " before renderpass " << pass.data.name << " because pass " << previous_pass.data.name << " writes to it as well";
-                        write_texture_barrier_necessity[prev_write_tex.name] = barrier_necessity::yes;
-                    }
-                }
-            }
-
-            if(idx == 0) {
-                break;
-            } else {
-                idx--;
-            }
-        }
-
-        for(const auto& [tex_name, _] : read_texture_barrier_necessity) {
-            const vk_texture& tex = textures.at(tex_name);
-
-            VkImageMemoryBarrier barrier = {};
-            barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-            barrier.srcAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-            barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-            barrier.oldLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-            barrier.newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-            barrier.srcQueueFamilyIndex = graphics_family_index;
-            barrier.dstQueueFamilyIndex = graphics_family_index;
-            barrier.image = tex.image;
-            barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-            barrier.subresourceRange.baseMipLevel = 0;
-            barrier.subresourceRange.levelCount = 1;
-            barrier.subresourceRange.baseArrayLayer = 0;
-            barrier.subresourceRange.layerCount = 1;
-
-            pass.read_texture_barriers.push_back(barrier);
-        }
-
-        for(const auto& [tex_name, _] : write_texture_barrier_necessity) {
-            if(tex_name == "Backbuffer") {
-                continue;
-            }
-
-            const VkImage image = textures.at(tex_name).image;
-
-            VkImageMemoryBarrier barrier = {};
-            barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-            barrier.srcAccessMask = VK_ACCESS_SHADER_READ_BIT;
-            barrier.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-            barrier.oldLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-            barrier.newLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-            barrier.srcQueueFamilyIndex = graphics_family_index;
-            barrier.dstQueueFamilyIndex = graphics_family_index;
-            barrier.image = image;
-            barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-            barrier.subresourceRange.baseMipLevel = 0;
-            barrier.subresourceRange.levelCount = 1;
-            barrier.subresourceRange.baseArrayLayer = 0;
-            barrier.subresourceRange.layerCount = 1;
-
-            pass.write_texture_barriers.push_back(barrier);
-        }
-    }
-
-    void vulkan_render_engine::generate_barriers_for_dynamic_resources() {
-        for(auto& [name, pass] : render_passes) {
-            create_barriers_for_renderpass(pass);
-        }
-    }
-
-    void vulkan_render_engine::update_material_descriptor_sets(const material_pass& mat, const std::unordered_map<std::string, vk_resource_binding>& name_to_descriptor) {
-        // for each resource:
-        //  - Get its set and binding from the pipeline
-        //  - Update its descriptor set
-        NOVA_LOG(TRACE) << "Updating descriptors for material " << mat.material_name;
-
-        std::vector<VkWriteDescriptorSet> writes;
-
-        // We create VkDescriptorImageInfo objects in a different scope, so were they to live there forever they'd get destructed before we can use them
-        // Instead we have them in a std::vector so they get deallocated _after_ being used
-        std::vector<VkDescriptorImageInfo> image_infos(mat.bindings.size());
-
-        std::vector<VkDescriptorBufferInfo> buffer_infos(mat.bindings.size());
-
-        for(const auto& [renderpass_name, pipelines] : pipelines_by_renderpass) {
-            bool should_break = false;
-
-            for(const vk_pipeline& pipeline : pipelines) {
-                if(pipeline.data.name == mat.pipeline) {
-                    for(const auto& [descriptor_name, resource_name] : mat.bindings) {
-                        if(pipeline.bindings.find(descriptor_name) == pipeline.bindings.end()) {
-                            NOVA_LOG(DEBUG) << "Material pass " << mat.name << " in material " << mat.material_name << " wants to bind " << resource_name << " to descriptor set " << descriptor_name << ", but it doesn't exist in pipeline "
-                                            << pipeline.data.name << ", which this material pass uses";
-                        }
-                    }
-
-                    should_break = true;
-                    break;
-                }
-
-                if(should_break) {
-                    break;
-                }
-            }
-        }
-
-        for(const auto& [descriptor_name, resource_name] : mat.bindings) {
-            const auto& descriptor_info = name_to_descriptor.at(descriptor_name);
-            const auto descriptor_set = mat.descriptor_sets[descriptor_info.set];
-            bool is_known = true;
-
-            VkWriteDescriptorSet write = {};
-            write.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            write.dstSet = descriptor_set;
-            write.dstBinding = descriptor_info.binding;
-            write.descriptorCount = 1;
-            write.dstArrayElement = 0;
-
-            if(textures.find(resource_name) != textures.end()) {
-                const vk_texture& texture = textures.at(resource_name);
-                write_texture_to_descriptor(texture, write, image_infos);
-
-            } else if(buffers.find(resource_name) != buffers.end()) {
-                const vk_buffer& buffer = buffers.at(resource_name);
-                write_buffer_to_descriptor(buffer, write, buffer_infos);
-
-            } else {
-                is_known = false;
-                NOVA_LOG(WARN) << "Resource " << resource_name << " is not known to Nova. I hope you aren't using it cause it doesn't exist";
-            }
-
-            if(is_known) {
-                writes.push_back(write);
-            }
-        }
-
-        vkUpdateDescriptorSets(device, static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
     }
 
     void vulkan_render_engine::write_texture_to_descriptor(const vk_texture& texture, VkWriteDescriptorSet& write, std::vector<VkDescriptorImageInfo>& image_infos) const {
